requirements.txt
Django >= 1.3, < 1.4
django-annoying >= 0.7.6
South==0.8.3
django-request==1.0
